Fix image upload
There was a problem with detecting the image size.
Sometimes the browser did not get the file type correctly.
Now, the mime-type is passed to and used inside function.

Also there was another issue related with mime-types when showing
images. The Matrix content API does sometimes not return the correcnt
mime-type. As a solution image elements do now also save their
mime-type.

Signed-off-by: Michael Weimann <michael.weimann@nordeck.net>