Fix duplicate text on paste
Before this change, text fields always start in edit mode. This
commit changes the behaviour to only go to edit mode if the text
field is empty (usually the case if a new text field is placed).
In addition to that the paste event for text fields is only handled,
if the text field is in edit mode (there is a cursor).

Signed-off-by: Michael Weimann <michael.weimann@nordeck.net>